copilot/requirements.txt
chainlit>=2.11.1
openai>=1.0.0
pyjwt>=2.8.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
